# Remove race condition when accessing remote bulker map