@cmacknz 2024-12-06
My take on this is that this is a real bug, the `GetBulkerMap()` method introduced to support remote outputs isn't concurrency safe and really can't be unless you do something like this. 

I don't like the use of `sync.Map` to solve this, primarily because it removes type safety in a codebase that I already find hard to follow with many levels of interface abstraction. I also don't see us satisfying the two criteria where sync.Map recommends its use: https://pkg.go.dev/sync#Map. I also don't like that it will encourage us to just randomly access the bulker everywhere but TBH that is kind of how it is designed to be used right now.

> The Map type is optimized for two common use cases: (1) when the entry for a given key is only ever written once but read many times, as in caches that only grow, or (2) when multiple goroutines read, write, and overwrite entries for disjoint sets of keys. In these two cases, use of a Map may significantly reduce lock contention compared to a Go map paired with a separate [Mutex](https://pkg.go.dev/sync#Mutex) or [RWMutex](https://pkg.go.dev/sync#RWMutex).

We don't satisfy 1 because outputs can be changed, updated, and deleted. We don't satisfy 2 because the entire point of `GetBulkMap` is to look at every key and not some non-overlapping set of keys.

It looks like we could solve this but just introducing a RWMutex for all the uses of bulkMap. There are several that are not mutex protected. https://github.com/elastic/fleet-server/blob/6b29ab474de9bbcbef890da5aeeb07737ffac0d9/internal/pkg/bulk/engine.go#L95


The `GetBulkerMap` method only has one use and it is to get the output name and client for each existing output. Rather than return a reference to the map, you can just hold a mutex to return a copy of those two things. Assuming the client is safe for concurrent access. `GetBulkerMap` shouldn't exist in its current form.

You could potentially rewrite the way the remote ES output healthcheck works to not need concurrency at all. The self monitor could listen on a channel for state updates from each attempt to interact with the remote ES update that could fail or something like that.