@michel-laterman 2024-12-06
I'll reimplement the mutex and create an issue to discuss remote output health