@michel-laterman 2024-12-06
Issue: https://github.com/elastic/fleet-server/issues/4185
I think we should remove remote health reporting from the policy self-monitor as it does nothing to the (fleet-server) status, but we can discuss it in the issue.